build: add standalone CLI release artifacts
