+++ backend/wails_app.go
@@ -1,3 +1,5 @@
+//go:build !cli
+
 package backend
 
 import (
