+++ cli_name.go
@@ -1,0 +1,8 @@
+//go:build !cli
+
+package main
+
+const (
+	cliExecutableName = "gotohp"
+	cliHasGUI         = true
+)
