+++ cli_name_cli.go
@@ -1,0 +1,8 @@
+//go:build cli
+
+package main
+
+const (
+	cliExecutableName = "gotohp-cli"
+	cliHasGUI         = false
+)
